# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_nate_go-boilerplate_pkg_response.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
...
  internal_http_handler.CreateNoteRequest:
    properties:
      content:
        example: This is the content of my new note
        maxLength: 10000
        type: string
      title:
        example: My New Note
        maxLength: 255
        minLength: 1
        type: string
    required:
    - title
    type: object
...
  internal_http_handler.NoteListResponse:
    properties:
      notes:
        items:
          $ref: '#/definitions/internal_http_handler.NoteResponse'
        type: array
      page:
        example: 1
        type: integer
      per_page:
        example: 20
        type: integer
      total:
        example: 42
        type: integer
      total_pages:
        example: 3
        type: integer
    type: object
  internal_http_handler.NoteResponse:
    properties:
      content:
        example: This is the content of my note
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      deleted_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      title:
        example: My Note
        type: string
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
    type: object
...
  internal_http_handler.UpdateNoteRequest:
    properties:
      content:
        example: This is the updated content
        maxLength: 10000
        type: string
      title:
        example: Updated Note Title
        maxLength: 255
        minLength: 1
        type: string
    required:
    - title
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/nate/go-boilerplate
  description: |-
    A production-ready Go REST API boilerplate with authentication, notes, and user management.
    Features include JWT authentication, rate limiting, soft deletes, and Redis-based token storage.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go Boilerplate API
  version: "1.0"
paths:
...
  /api/v1/notes:
    get:
      consumes:
      - application/json
      description: Get a paginated list of notes belonging to the authenticated user.
        Supports soft-deleted notes with with_deleted parameter.
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 100
        minimum: 1
        name: per_page
        type: integer
      - default: false
        description: Include soft-deleted notes
        in: query
        name: with_deleted
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of notes
          schema:
            $ref: '#/definitions/internal_http_handler.NoteListResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user's notes
      tags:
      - notes
    post:
      consumes:
      - application/json
      description: Create a new note for the authenticated user.
      parameters:
      - description: Note details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.CreateNoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Note created successfully
          schema:
            $ref: '#/definitions/internal_http_handler.NoteResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new note
      tags:
      - notes
  /api/v1/notes/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a note (move to trash). User must own the note.
      parameters:
      - description: Note ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Note deleted successfully
        "400":
          description: Invalid note ID format
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "404":
          description: Note not found
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Soft delete a note
      tags:
      - notes
    get:
      consumes:
      - application/json
      description: Retrieve a specific note by its ID. User must own the note.
      parameters:
      - description: Note ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Note details
          schema:
            $ref: '#/definitions/internal_http_handler.NoteResponse'
        "400":
          description: Invalid note ID format
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "404":
          description: Note not found
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a note by ID
      tags:
      - notes
    put:
      consumes:
      - application/json
      description: Update an existing note. User must own the note.
      parameters:
      - description: Note ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Updated note details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.UpdateNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Note updated successfully
          schema:
            $ref: '#/definitions/internal_http_handler.NoteResponse'
        "400":
          description: Invalid request body, note ID, or validation error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "404":
          description: Note not found
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a note
      tags:
      - notes
  /api/v1/notes/{id}/restore:
    post:
      consumes:
      - application/json
      description: Restore a soft-deleted note from trash. User must own the note.
      parameters:
      - description: Note ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Note restored successfully
        "400":
          description: Invalid note ID format
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "404":
          description: Note not found or not deleted
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore a soft-deleted note
      tags:
      - notes
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"